Set up grafzahl to be used on Google Colab or similar environments

Set up grafzahl to be used on Google Colab or similar environments. This function is also useful if you do not want to use conda on a local machine, e.g. you have configurateed the required Python package.

Usage

use_nonconda(install = TRUE, check = TRUE, verbose = TRUE)

Arguments

install

logical, whether to install the required Python packages

check

logical, whether to perform a check after the setup. The check displays 1) whether CUDA can be detected, 2) whether the non-conda mode has been activated, i.e. whether the option 'grafzahl.nonconda' is TRUE.

verbose,

logical, whether to display messages

Value

TRUE (invisibly) if installation is successful.

Examples

# A typical use case for Google Colab
if (interactive()) {
    use_nonconda()
}